+++ dwds/lib/src/debugging/metadata/function.dart
@@ -18,24 +18,17 @@ class FunctionMetaData {
     RemoteObject remoteObject,
   ) async {
     final evalExpression = '''
-      function(remoteObject) {
-        var sdkUtils = ${globalLoadStrategy.loadModuleSnippet}('dart_sdk').dart;
-        var name = remoteObject.name;
-        if(remoteObject._boundObject) {
-         name = sdkUtils.getType(remoteObject._boundObject).name +
-                 '.' + remoteObject._boundMethod.name;
-        }
-        return name;
+      function() {
+        const sdk = ${globalLoadStrategy.loadModuleSnippet}('dart_sdk');
+        const dart = sdk.dart;
+        return dart.getFunctionMetadata(this);
       }
     ''';
-    final arguments = [
-      {'objectId': remoteObject.objectId},
-    ];
+
     final response = await remoteDebugger.sendCommand(
       'Runtime.callFunctionOn',
       params: {
         'functionDeclaration': evalExpression,
-        'arguments': arguments,
         'objectId': remoteObject.objectId,
         'returnByValue': true,
       },
