+++ fixtures/_webdevSoundSmoke/build.yaml
@@ -1,7 +1,7 @@
-# global_options:
-#   build_web_compilers:ddc:
-#     options:
-#       canary: false
-#   build_web_compilers:sdk_js:
-#     options:
-#       canary: false
+global_options:
+  build_web_compilers:ddc:
+    options:
+      canary: true
+  build_web_compilers:sdk_js:
+    options:
+      canary: true
